Fixed Issue with Input not being shown, display now prints out balance, moved some tests for Bank.java to a new file called BankTest.java
